requirements.txt
numpy<2.0.0
streamlit
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-chroma
langchain-google-genai
chromadb
sentence-transformers
PyPDF2
python-docx
python-dotenv
